refactor: consolidate the tap window-global seams into TapManagerDataSource
